# contracts/http/openapi.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "EcoBin P0 HTTP API",
    "version": "1.0.0-rc.1",
    "description": "EcoBin /api/v1 transport contract. Business slices add their own paths and schemas without redefining these cross-cutting rules."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment"
    }
  ],
...
  "paths": {
...
    "/api/v1/web/platform/tenants": {
      "get": {
        "operationId": "listIdentityTenants",
        "tags": ["Identity directory"],
        "summary": "List tenants as a platform administrator",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {"$ref": "#/components/parameters/DirectoryStatusFilter"},
          {"$ref": "#/components/parameters/DirectoryQuery"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"}
        }
      },
      "post": {
        "operationId": "createIdentityTenant",
        "tags": ["Identity directory"],
        "summary": "Create one disabled tenant",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateTenantRequest"}}}
        },
        "responses": {
          "201": {"$ref": "#/components/responses/DirectoryCreated"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}": {
      "parameters": [{"$ref": "#/components/parameters/TenantCode"}],
      "get": {
        "operationId": "getIdentityTenant",
        "tags": ["Identity directory"],
        "summary": "Read one tenant through the platform privilege boundary",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/profile": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "put": {
        "operationId": "updateIdentityTenantProfile",
        "tags": ["Identity directory"],
        "summary": "Replace tenant profile with optimistic concurrency",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TenantProfileRequest"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/principal-account": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "post": {
        "operationId": "createTenantPrincipalAccount",
        "tags": ["Identity directory"],
        "summary": "Create the protected tenant-principal account",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreatePrincipalAccountRequest"}}}
        },
        "responses": {
          "201": {"$ref": "#/components/responses/DirectoryCreated"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/principal-account/password-resets": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "post": {
        "operationId": "resetTenantPrincipalPassword",
        "tags": ["Identity directory"],
        "summary": "Reset the protected tenant-principal password and revoke sessions",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResetPasswordRequest"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/activations": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "post": {
        "operationId": "activateIdentityTenant",
        "tags": ["Identity directory"],
        "summary": "Enable a tenant after its principal exists",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/deactivations": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "post": {
        "operationId": "deactivateIdentityTenant",
        "tags": ["Identity directory"],
        "summary": "Disable a tenant and revoke every tenant session",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
...
    "/api/v1/web/platform/tenants/{tenantCode}/organizations": {
      "parameters": [{"$ref": "#/components/parameters/TenantCode"}],
      "get": {
        "operationId": "listPlatformIdentityOrganizations",
        "tags": ["Identity directory"],
        "summary": "List organizations in one explicit platform target tenant",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {"$ref": "#/components/parameters/DirectoryStatusFilter"},
          {"$ref": "#/components/parameters/DirectoryQuery"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      },
      "post": {
        "operationId": "createPlatformIdentityOrganization",
        "tags": ["Identity directory"],
        "summary": "Create a disabled organization in one explicit platform target tenant",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateOrganizationRequest"}}}
        },
        "responses": {
          "201": {"$ref": "#/components/responses/DirectoryCreated"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/OrganizationCode"}
      ],
      "get": {
        "operationId": "getPlatformIdentityOrganization",
        "tags": ["Identity directory"],
        "summary": "Read an organization in one explicit platform target tenant",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/profile": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/OrganizationCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "put": {
        "operationId": "updatePlatformIdentityOrganizationProfile",
        "tags": ["Identity directory"],
        "summary": "Replace an organization profile in an explicit tenant",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrganizationProfileRequest"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/activations": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/OrganizationCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "post": {
        "operationId": "activatePlatformIdentityOrganization",
        "tags": ["Identity directory"],
        "summary": "Activate an organization in an explicit tenant",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/deactivations": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/OrganizationCode"},
        {"$ref": "#/components/parameters/IdempotencyKey"}
      ],
      "post": {
        "operationId": "deactivatePlatformIdentityOrganization",
        "tags": ["Identity directory"],
        "summary": "Deactivate an organization and remove its live access",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
...
    "/api/v1/web/platform/tenants/{tenantCode}/staff-accounts": {
      "parameters": [{"$ref": "#/components/parameters/TenantCode"}],
      "get": {
        "operationId": "listPlatformStaffAccounts",
        "tags": ["Identity directory"],
        "summary": "List staff in one explicit platform target tenant",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {"$ref": "#/components/parameters/DirectoryStatusFilter"},
          {"$ref": "#/components/parameters/DirectoryQuery"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      },
      "post": {
        "operationId": "createPlatformStaffAccount",
        "tags": ["Identity directory"],
        "summary": "Create tenant-scoped staff through a platform privilege use case",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateStaffAccountRequest"}}}
        },
        "responses": {
          "201": {"$ref": "#/components/responses/DirectoryCreated"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
...
    "/api/v1/web/platform/tenants/{tenantCode}/staff-accounts/{staffUid}": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/StaffAccountUid"}
      ],
      "get": {
        "operationId": "getPlatformStaffAccount",
        "tags": ["Identity directory"],
        "summary": "Read staff only inside one explicit tenant",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "webSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-ecobin-web-session",
        "description": "Secure, HttpOnly, SameSite=Lax, Path=/ and no Domain. JavaScript never reads this value."
      },
      "webCsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN",
        "description": "Bootstrapped from GET /api/v1/web/auth/csrf-token and rotated after login/logout."
      },
      "miniappBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Exactly one short-lived aud=miniapp or aud=miniapp-staff session. No refresh token."
      },
      "wechatPaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Signature"
      },
      "wechatPayTimestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Timestamp"
      },
      "wechatPayNonce": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Nonce"
      },
      "wechatPaySerial": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Serial"
      }
    },
...
  },
...
}